Refactor/1.0 (#22)
* Changed Service and DAO calls

* Changed Models and Storage Functions

* Fixed Typo in test
