+++ backend/controllers/status.ts
@@ -1,9 +1,9 @@
 import express from "express";
-import { getCurrentStatus } from "../services/status";
+import * as statusService from "../services/status";
 const router = express.Router();
 
 router.get('/', (req, res) => {
-    getCurrentStatus().then(status => {
+    statusService.getCurrentStatus().then(status => {
         res.status(200).json(status);
     }).catch(err => {
         res.status(500).send();
