+++ backend/services/auth.ts
@@ -1,9 +1,10 @@
 import * as bcrypt from 'bcrypt';
 import * as jwt from 'jsonwebtoken';
-import { getUserByEmail, UserLogin } from "../models/user";
+import * as blacklistDAO from '../storage/blacklist';
+import * as userDAO from "../storage/user";
 import { getSecret } from '../util/secret';
+import { UserLogin } from '../models/user';
 import { jwtContentTransformer } from '../util/transformer';
-import { addJWTToBlacklist } from '../models/blacklist';
 
 export interface JWTContent {
     id: string | undefined;
@@ -13,7 +14,7 @@ export interface JWTContent {
 export function login(userCredentials: UserLogin): Promise<{}> {
     return new Promise((resolve, reject) => {
         //Get User from DB to check Credentials
-        getUserByEmail(userCredentials.email).then(user => {
+        userDAO.getUserByEmail(userCredentials.email).then(user => {
             //BCrypt Current Password
             bcrypt.compare(userCredentials.password, user.password).then(result => {
                 if (result) {
@@ -38,6 +39,6 @@ export function login(userCredentials: UserLogin): Promise<{}> {
 export function logout(jwtKey: string): Promise<{}> {
     return new Promise((resolve, reject) => {
         if (!jwtKey) {reject(new Error("No JWT Key"))};
-        addJWTToBlacklist(jwtKey).then(() => resolve({ status: 'success' })).catch(() => reject(new Error("Server Error")));
+        blacklistDAO.addJWTToBlacklist(jwtKey).then(() => resolve({ status: 'success' })).catch(() => reject(new Error("Server Error")));
     });
 }
